feat(smc): ecm重构
使用groovy组件重构ECM
