+++ smc/src/main/java/com/tlcsdm/smc/codeDev/ecm/AbstractU2XFamilyScript.java
@@ -94,7 +94,7 @@ public abstract class AbstractU2XFamilyScript extends AbstractEcmScript {
             operation.put("funcId", funcId);
             operation.put("support", String.valueOf(support));
             operation.put("errorNote", "");
-            // 数据后置处理
+            // function数据后置处理
             handlerOperationSupport(operation, funcSupCondition, optMaskintStatus);
             function.add(operation);
         }
@@ -116,102 +116,11 @@ public abstract class AbstractU2XFamilyScript extends AbstractEcmScript {
     /**
      * errorSource 数据后续处理
      */
-    protected void handlerErrorSourceMap(Map<String, Object> errorSource, String product, int optErrortIndex) {
-        String errorSourceenName = (String) errorSource.get("errorSourceEnName");
-        String errorSourcejpName = (String) errorSource.get("errorSourceJpName");
-        errorSourceenName = cleanErrorSourceData(errorSourceenName);
-        errorSourcejpName = cleanErrorSourceData(errorSourcejpName);
-        if (errorSourceenName.endsWith("*7")) {
-            errorSourceenName = StrUtil.replaceLast(errorSourceenName, "*7", "(For debug purpose only)");
-            if (errorSourcejpName.endsWith("*7")) {
-                errorSourcejpName = StrUtil.replaceLast(errorSourcejpName, "*7", "(デバッグ目的のみ)");
-            } else {
-                errorSourcejpName += "(デバッグ目的のみ)";
-            }
-        }
-        errorSource.put("errorSourceEnName", errorSourceenName);
-        errorSource.put("errorSourceJpName", errorSourcejpName);
-
-        List<Map<String, Object>> function = (List<Map<String, Object>>) errorSource.get("function");
-        List<Map<String, Object>> extraFunc = new ArrayList<>();
-        for (Map<String, Object> map : function) {
-            String funcId = map.get("funcId").toString();
-            if ("optErrort".equals(funcId)) {
-                String support = map.get("support").toString();
-                String errorNote = map.get("errorNote").toString();
-                int size = 0;
-                if (product.startsWith("RH850U2A16")) {
-                    size = 4;
-                    generateErrort(size, support, errorNote, extraFunc, function);
-                    function.remove(map);
-                } else if (product.startsWith("RH850U2A8") || product.startsWith("RH850U2A6")) {
-                    size = 2;
-                    generateErrort(size, support, errorNote, extraFunc, function);
-                    function.remove(map);
-                } else {
-                    break;
-                }
-            }
-        }
-        function.addAll(optErrortIndex, extraFunc);
-    }
-
-    private void generateErrort(int size, String support, String errorNote, List<Map<String, Object>> extraFunc,
-        List<Map<String, Object>> function) {
-        for (int i = 0; i < size; i++) {
-            Map<String, Object> map = new HashMap<>();
-            map.put("funcId", "optErrort" + i);
-            map.put("support", support);
-            map.put("errorNote", errorNote);
-            extraFunc.add(map);
-        }
-    }
-
-    protected void handlerOperationSupport(Map<String, Object> operation, String funcSupCondition,
-        boolean optMaskintStatus) {
-        if (funcSupCondition.contains("*")) {
-            String mesNum = StrUtil.subAfter(funcSupCondition, "*", true);
-            if ("1".equals(mesNum) || "2".equals(mesNum)) {
-                operation.put("errorNote", mesNum);
-            }
-            if ("5".equals(mesNum)) {
-                String funcId = operation.get("funcId").toString();
-                if ("optDCLS".equals(funcId)) {
-                    operation.put("support", String.valueOf(optMaskintStatus));
-                }
-                if ("optIntg".equals(funcId)) {
-                    operation.put("support", "false");
-                }
-            }
-        } else {
-            String funcId = operation.get("funcId").toString();
-            if ("optDCLS".equals(funcId)) {
-                operation.put("support", "false");
-            }
-            if ("optIntg".equals(funcId)) {
-                operation.put("support", String.valueOf(optMaskintStatus));
-            }
-        }
-
-    }
+    protected abstract void handlerErrorSourceMap(Map<String, Object> errorSource, String product, int optErrortIndex);
 
     /**
-     * 清洗ErrorSource数据
+     * function数据后置处理
      */
-    private String cleanErrorSourceData(String data) {
-        data = data.replaceAll("  ", " ");
-        if (data.contains(" (")) {
-            data = StrUtil.replace(data, " (", "(");
-        }
-        if (data.contains("\n")) {
-            List<String> list = StrUtil.split(data, "\n");
-            data = list.get(0);
-            for (int i = 1; i < list.size(); i++) {
-                data += " ";
-                data += list.get(i);
-            }
-        }
-        data = data.replaceAll("  ", " ");
-        return data;
-    }
+    protected abstract void handlerOperationSupport(Map<String, Object> operation, String funcSupCondition,
+                                                    boolean optMaskintStatus);
 }
