--- smc/src/main/resources/com/tlcsdm/smc/static/groovy/codeDev/ecm/test.groovy
@@ -1,12 +1,0 @@
-def hello() {
-    println "Hello QE"
-    demo()
-    return "success, hello QE"
-}
-
-def demo() {
-    println "Hello demo"
-}
-
-hello()
-println "Hello, welcome to Groovy!"
